alterado meses

+++ CalculoSoft/Communs/SoftRule.cs
@@ -8,7 +8,7 @@ namespace CalculoSoft.Communs
     public static class SoftRule
     {
         //Esse método irá chamar a api aonde fica a nossa taxa 
-        public static async Task<double> RetornaCalculo(double vlrIni, double meses, string url, bool teste = false, double?valor = null)
+        public static async Task<double> RetornaCalculo(double vlrIni, int meses, string url, bool teste = false, double?valor = null)
         {
             double taxa = 0;
             //valida testes passando a taxa como parametro
